weather-api/requirements.txt
Flask
requests
redis
python-dotenv
Flask-Limiter
geopy
